// src/modules/meiko/enums/status.enum.ts
export enum InvoiceStatus {
  PROCESADO = 1,
  NO_APLICA_PARA_EL_ESTUDIO = 2,
  NO_PROCESABLE = 3,
  GENERACION_PLANTILLA = 4,
  ERROR_DE_DESCARGA = 5,
  FECHA_NO_VALIDA = 6,
}

export enum DeliveryStatus {
  PROCESADO = 'PROCESADO',
  NO_APLICA_PARA_EL_ESTUDIO = 'NO_APLICA_PARA_EL_ESTUDIO',
  NO_PROCESABLE = 'NO_PROCESABLE',
  GENERACION_PLANTILLA = 'GENERACION_PLANTILLA',
  ERROR_DE_DESCARGA = 'ERROR_DE_DESCARGA',
  FECHA_NO_VALIDA = 'FECHA_NO_VALIDA',
}